+++ dify_studio/kaye_chat/nodes/post_start.py
@@ -7,7 +7,7 @@ OUTPUT_SKIP_KEY = "skip_sense"
 
 
 # helpers  #####################################################################
-def should_skip_sense(role, llm_override, difficulty_override):
+def should_skip_sense(role, llm_override):
     """
     decide whether should skip sense node for this round of conversation
     based on combination of difficulty & LLM overrides given
@@ -17,18 +17,10 @@ def should_skip_sense(role, llm_override, difficulty_override):
     :type role: str
     :param llm_override:
     :type llm_override: str
-    :param difficulty_override:
-    :type difficulty_override: float
     :return: whether should skip sense node
     :rtype: bool
     """
-    # Bug dont skip for coder, b/c still need to extract PLs during sensing
-
-    if role == "coder":
-        # skip for kyc, when difficult is provided
-        return bool(difficulty_override)
-
-    elif role == "barista":
+    if role == "barista":
         # skip for Assistant Barista
         return True
 
@@ -47,7 +39,6 @@ def should_skip_sense(role, llm_override, difficulty_override):
 def main(
     role_override: str,
     llm_override: str,
-    difficulty_override: float,
     current_role: str,
 ):
     """
@@ -59,8 +50,6 @@ def main(
     :type role_override: str
     :param llm_override:
     :type llm_override: str
-    :param difficulty_override:
-    :type difficulty_override: float
     :param current_role: role saved in conversation variable,
             often decided on 1st round of conversation
     :type current_role: str
@@ -72,6 +61,6 @@ def main(
     # decide role  =============================================================
     role = role_override or current_role or ""
 
-    skip_sense = should_skip_sense(role, llm_override, difficulty_override)
+    skip_sense = should_skip_sense(role, llm_override)
 
     return {OUTPUT_ROLE_KEY: role, OUTPUT_SKIP_KEY: skip_sense}